cli!: tidy things into sub-commands
Tidy the CLI code, and add sub-commands.

As a consequence of this latter change, to run conjure oxide, one must
now do `cargo run -- solve <input>` or `conjure_oxide solve <input>`.
This is because I could not find a good way to do default arguments with
clap.

Other tidying includes:

* Move CLI flags into a separate file
* Move logging setup into its own function
* Move things for each subcommand into their own files.

The motivation of this work is to allow more subcommands to be added in
the future.

Before this change, we had the `--print-info-schema` flag that ignored
all other flags, just printed the schema, and exited. Using proper
subcommands is a more scalable way to add features like this, with the
added advantage that subcommands can have their own arguments.
