+++ conjure_oxide/src/help_text.txt
@@ -24,9 +24,9 @@ LOGGING RULE TRACES
 For example, to see TRACE logs in a pretty format (mainly useful for
 debugging):
 
-    $ RUST_LOG=trace conjure-oxide --verbose <model>
+    $ RUST_LOG=trace conjure-oxide solve --verbose <model>
 
     # or through cargo:
-    $ RUST_LOG=trace cargo run -- --verbose <model>
+    $ RUST_LOG=trace cargo run -- solve --verbose <model>
